Add Toolbar with Image Picker
Add a toolbar to the bottom of the View Controller view with a single button to open an image picker. The delegate method for UIImagePickerController has been implemented but does not yet return the selected image to the view for display. Text is printed to the debug console to verify that the delegate methods for selecting an image and dismissing the photo picker are properly implemented.