Add Text Fields to View
Add two text fields to the view: topTextField and bottomTextField. Both text fields are configured using the custom function formatTextField(_:, defaultText:). This function sets the delegate of the UITextField to self, centers the text in the text field and assigns default text if provided.

UITextField Delegate functions textFieldDidBeginEditing(_:) and textFieldShouldReturn(_:) are implemented to clear the text field when editing begins and hide the keyboard when editing is finished.