Change Text Appearance
Change the appearance of the text in the view to more closely resemble what is seen in memes. This involves changing the attributes of the UITextFields for capitalization and border style on the storyboard, as well as using an array of NSAttributedString keys to modify the default text attributes at runtime.