Navigation, Toolbar Changes and Generate and Share Meme Image
This commit embeds the view controller inside a navigation controller, which includes both the navigatoion bar as well as a toolbar. This change warranted the removal of the toolbar object that was initially placed, moving the existing buttons to the new embedded toolbar. The navigation bar and toolbar are automatically hidden when the view is rotated into landscape view.

Add: Function generateMemeImage() which in quick succession, hides the navigation and toolbars, renders the view to an image, and then shows the navigation and toolbars.

Add: Function saveMeme which generates a Meme object containing the top and bottom text strings, the original image file, as well as the newly generated Meme image.

Change: Changed the autolayout constraints of the imageView so that the view is always square in both landscape and portrait.

Fix: Changed selectedTextField to an optional because the application was crashing when the activity view pushed the user into a different application.