Save Meme to Device
Add/Fix: Add NSPhotoLibraryAddUsageDescription to Info.plist to gain permission to save images to the user's device. This fixes a bug where the app would crash if the user attempted to save the image to the device from the activity controller.

Add: completion handler added to activity view controller to generate the Meme object after the meme has been shared or saved to disk. The completion handler is currently calling the saveMeme function but does not do anything with this object at the moment.