Change UIImageView Constraints
This commit changes the constraints and content mode of the UIImageView to meet project specifications outlined by Udacity. Unfortunately the results are less than ideal aesthetically as the text is not always placed on top of the image and is sometimes in the margins outside the image and the text is not constrained to the width of the image.