Clear Text Fields Only If They Contain Placeholder Text
Change: The top or bottom text field will only be cleared if it contains the placeholder text. This prevents custom strings from being cleared if the user only wishes to update the text. This was accomplished by adding a struct with string constants for the top and bottom placeholders, determining which text field is selected and comparing the text with the placeholder text.

The placeholder text constants are also used to set the initial values of the text fields.