Add Additional Navigation Structure
This commit adds a tab view that will be used to switch between a table/list view and a collection/grid view of sent memes. In the upper-right of the display is a "+" button that displays the "Create Meme" view that has been the intial view up until this point. Finally, a "cancel" button has been added to the "Create Meme" view to dismiss the view and return to either the table or collection view.