Add Sent Memes Table View Prototype
This commit adds a Table View prototype for displaying a list of sent memes. Each cell contains one image that will serve as the meme thumbnail and two labels: top text label and bottom text label. Tapping a cell opens the editMemeViewController and changes the title of the view Controller to "Edit Meme", rather than the default "Create Meme."

Currently the data in the table is static with a fixed number of rows, but will ultimately read the memes array from the app delegate.