Save and Edit Memes and Display in Sent Memes Table
This commit is centered around saving new memes to the memes array in the ap delegate and editing said memes and resaving them to the array.

After a meme is created the meme object is added to the memes array in the app delegate. This memes array is then used to populate the Sent Memes Table View. Tapping on any of the existing memes dispays a new view called ViewMemeViewController. From this view the user can view the meme image, tapping on the image to hide/show the navigation bar. From the navigation bar the user can choose to return to the previous screen (table view) or edit the meme, which passes the meme object and populates the text fields and image view in the EditMemeViewController (this same view is used to create new memes). If the user cancels editing the meme they are brought back to the view meme view. If the user saves the meme by making changes, tapping the share button and then choosing an activity. Clicking done from the activity view will then return the user to the root view/table view.

Because the view was not popping to the root view from the completion handler of the dismiss function, I handled this by using the notification center, posting and listening for a custom notification called didFinishEditing.