Display Memes in Collection View
Display memes from the shared memes array found in the app delegate in a UICollectionView. Selecting a meme removes the meme from the array and sends it to the view meme view controller where it will either be reinserted into the array if the user returns to the collection view without editing, or will be replaced with an edited version if the user chooses to edit the meme.