Add Ability to Change Fonts
This commit adds the ability to change the font being used in the meme editor. Tapping on the A icon in the toolbar of the meme editor beings up an action sheet with 3 font choices -- the font that is currently being used is marked with a check.

In order to edit an existing meme and automatically select the correct font, a font property was added to the Meme struct.

This commit also adds the Impact font and sets it as the default font.