Fix top image constraint based on device orientation
These changes alter the image constraint depending on the device's orientation. Prior to these changes there would be a gap between the navigation bar and the image when the device was rotated to landscape. These changes check both the devices current orientation as well as the orientation of the status bar to eliminate false positives when tapping on the display when the device was lying flat.