Add edit tools pane with text positioning tool
This commit adds a tool pane to the EditMemeViewController the button to toggle this pane on/off replaces the toolbar button to change the font. The option to change the font was moved onto this tool pane. To facilitate this change the formatTextField(_:defaultText:font:) function was replaced by adding an extension method to UITextField called setFont(_:) that sets the font with the appropriate appearance attributes. In addition to the change font tool is the addition of a text positioning tool. The text positioning tool allows the use to change the vertical positoning of the top and bottom text fields. The behavior of this tool moves both text fields simultaneously.